+++ src/OpenTelemetry/Metrics/ExplicitBucketHistogramConfiguration.cs
@@ -14,14 +14,10 @@
 // limitations under the License.
 // </copyright>
 
-using System;
-
 namespace OpenTelemetry.Metrics
 {
     public class ExplicitBucketHistogramConfiguration : MetricStreamConfiguration
     {
-        private Aggregation aggregation = Aggregation.Histogram;
-
         /// <summary>
         /// Gets or sets the values representing explicit histogram bucket
         /// boundary values.
@@ -30,20 +26,5 @@ namespace OpenTelemetry.Metrics
         /// The array must be in ascending order with distinct values.
         /// </remarks>
         public double[] Boundaries { get; set; }
-
-        public override Aggregation Aggregation
-        {
-            get => this.aggregation;
-
-            set
-            {
-                if (value != Aggregation.Histogram)
-                {
-                    throw new ArgumentException($"Aggregation must be Histogram.");
-                }
-
-                this.aggregation = value;
-            }
-        }
     }
 }